Changed python to float32 instead of 64 for fair comparison

+++ StudyPython/StudyPython.py
@@ -228,8 +228,8 @@ def split_train_infer(
 
     # --- Split phase ---
     def _split():
-        X = sub[features].to_numpy(dtype=np.float64, copy=False)
-        y = sub[TARGET].to_numpy(dtype=np.float64, copy=False)
+        X = sub[features].to_numpy(dtype=np.float32, copy=False)
+        y = sub[TARGET].to_numpy(dtype=np.float32, copy=False)
         return train_test_split(X, y, test_size=test_size, random_state=seed)
 
     (X_train, X_test, y_train, _), split_ns, split_heap, split_rss = _measure(_split)
